RayTracing first commit

+++ GFlat.cc
@@ -21,10 +21,10 @@ GVector3 GFlat::getPoint() const {
 GVector3 GFlat::getNormal(const GVector3& v) const {
   return m_Normal;
 }
-INTERSECTION_TYPE GFlat::isIntersected(const Ray& ray,float& point) const {
-  float tmp1=m_Point*m_Normal - m_Normal*ray.getOrigin();
-  float tmp2=m_Normal*ray.getDirection();
-  float t=tmp1/tmp2;
+INTERSECTION_TYPE GFlat::isIntersected(const Ray& ray,double& point) const {
+  double tmp1=m_Point*m_Normal - m_Normal*ray.getOrigin();
+  double tmp2=m_Normal*ray.getDirection();
+  double t=tmp1/tmp2;
   if (t<0)
     return MISS;
   if (point>t) {
