+++ inc/GVector3.h
@@ -5,29 +5,29 @@
 
 class GVector3 {
   public:  
-    float x;
-    float y;
-    float z;
+    double x;
+    double y;
+    double z;
   public:
     friend std::ostream& operator<<(std::ostream&,const GVector3&);
     friend std::istream& operator>>(std::istream&,GVector3&);
     GVector3();
     GVector3(const GVector3&);
-    GVector3(float,float,float);
-    float length() const;
+    GVector3(double,double,double);
+    double length() const;
     //false if the length==0
     bool normalize();
     GVector3 operator=(const GVector3&);
-    GVector3 operator=(float);
+    GVector3 operator=(double);
     GVector3 operator+(const GVector3&) const;
     GVector3 operator+=(const GVector3&);
     GVector3 operator-(const GVector3&) const;
     GVector3 operator-=(const GVector3&);
-    GVector3 operator*(const float) const;
-    GVector3 operator*=(const float);
-    float operator*(const GVector3&) const;
-    GVector3 operator/(const float) const;
-    GVector3 operator/=(const float);
+    GVector3 operator*(const double) const;
+    GVector3 operator*=(const double);
+    double operator*(const GVector3&) const;
+    GVector3 operator/(const double) const;
+    GVector3 operator/=(const double);
 };
 std::ostream& operator<<(std::ostream&,const GVector3&);
 std::istream& operator>>(std::istream&,GVector3&);
